rearranged matcher code for reuse
