+++ Sources/Matchers/TypeMatcher.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+@objc
+public class TypeMatcher: NSObject {
+  let typeValue: Any
+
+  public init(value: Any) {
+    self.typeValue = value
+  }
+
+  @objc public func value() -> Any {
+    return typeValue
+  }
+}
