--- Sources/Native/Matchers/TypeMatcher.swift
@@ -1,18 +1,0 @@
-import Foundation
-
-@objc
-public class TypeMatcher: NSObject, MatchingRule {
-  let typeValue: Any
-
-  public init(value: Any) {
-    self.typeValue = value
-  }
-
-  public func rule() -> [String: Any] {
-    return [ "match": "type" ]
-  }
-
-  public func value() -> Any {
-    return typeValue
-  }
-}
