+++ Sources/Ruby/RubyMatcher.swift
@@ -18,17 +18,7 @@ open class RubyMatcher: NSObject, Matchers {
   }
 }
 
-// merge-todo clean up / split up files / maybe generalize
-@objc
-public class RubyTermMatcher: NSObject, MatchingRule {
-  let typeValue: String
-  let regex: String
-
-  public init(regex: String, value: String) {
-    self.typeValue = value
-    self.regex = regex
-  }
-
+public class RubyTermMatcher: TermMatcher, MatchingRule {
   public func rule() -> [String: Any] {
     return [ "json_class": "Pact::Term",
     "data": [
@@ -39,51 +29,23 @@ public class RubyTermMatcher: NSObject, MatchingRule {
         "s": regex]
     ] ]
   }
-
-  public func value() -> Any {
-    return typeValue
-  }
 }
 
-@objc
-public class RubyTypeMatcher: NSObject, MatchingRule {
-  let typeValue: Any
-
-  public init(value: Any) {
-    self.typeValue = value
-  }
-
+public class RubyTypeMatcher: TypeMatcher, MatchingRule {
   public func rule() -> [String: Any] {
     return [
       "json_class": "Pact::SomethingLike",
       "contents": typeValue
     ]
   }
-
-  public func value() -> Any {
-    return typeValue
-  }
 }
 
-@objc
-public class RubyMinTypeMatcher: NSObject, MatchingRule {
-    let typeValue: [String: Any]
-    let min: Int
-
-    public init(value: [String: Any], min: Int) {
-        self.typeValue = value
-        self.min = min
-    }
-
+public class RubyMinTypeMatcher: MinTypeMatcher, MatchingRule {
     public func rule() -> [String: Any] {
         return [
           "json_class": "Pact::ArrayLike",
           "contents": typeValue,
           "min": min
         ]
-    }
-
-    public func value() -> Any {
-        return typeValue
     }
 }
